+++ cnpj-receita-federal-brasil/visitor/br/gov/economia/receita/IRegisterVisitor.java
@@ -3,27 +3,27 @@ package br.gov.economia.receita;
 import br.gov.economia.receita.imp.VisitResult;
 
 public interface IRegisterVisitor {
-  public VisitResult beginHeader(int row);
-  public VisitResult fieldHeader(int row, IField field);
+  public VisitResult beginHeader(long row);
+  public VisitResult fieldHeader(long row, IField field);
   public VisitResult endHeader();
   
-  public VisitResult beginEmpresa(int row);
-  public VisitResult fieldEmpresa(int row, IField field);
+  public VisitResult beginEmpresa(long row);
+  public VisitResult fieldEmpresa(long row, IField field);
   public VisitResult endEmpresa();
   
-  public VisitResult beginSocio(int row);
-  public VisitResult fieldSocio(int row, IField field);
+  public VisitResult beginSocio(long row);
+  public VisitResult fieldSocio(long row, IField field);
   public VisitResult endSocio();
       
-  public VisitResult beginCnae(int row);
-  public VisitResult fieldCnae(int row, IField field);
+  public VisitResult beginCnae(long row);
+  public VisitResult fieldCnae(long row, IField field);
   public VisitResult endCnae();
 
-  public VisitResult beginTrailler(int row);
-  public VisitResult fieldTrailler(int row, IField field);
+  public VisitResult beginTrailler(long row);
+  public VisitResult fieldTrailler(long row, IField field);
   public VisitResult endTrailler();
   
-  public VisitResult handleError(int row, Throwable e);
+  public VisitResult handleError(long row, Throwable e);
   
   public void end();
   public void start();
